fix(Rest): use reference type for DOM
We had someone in Sapphire running into an issue here because they
didn't have "DOM" in their tsconfig lib, nor did they have
"skipLibCheck" set to true. While those 2 options are viable
end-user solutions, it would be better for the lib to actually
just function without the need for that.

This can be achieved with a simple triple slash directive, which unlike
the lib config in tsconfig.json does get copied over to the build
output `index.d.ts`